feat(preview): add route
- preview allows for post message events to change config w/o reload
- preview sends post message with embed url, html, height, and styles
- move embed page content to embed component
- add helper functions to generate embed height and html style attibs
- improve embed height calcs and styles
- update caniuse db
