+++ components/Player/PlayerShareMenu/PlayerShareMenu.tsx
@@ -21,6 +21,7 @@ export interface IPlayerShareMenuProps extends IModalProps {
   onOpen(): void;
   className?: string;
   embedHtml?: string;
+  isPrivate?: boolean;
 }
 
 const PlayerShareMenu: React.FC<IPlayerShareMenuProps> = ({
@@ -29,6 +30,7 @@ const PlayerShareMenu: React.FC<IPlayerShareMenuProps> = ({
   isOpen,
   portalId,
   embedHtml,
+  isPrivate,
   className
 }) => {
   const handleClick = () => {
@@ -55,16 +57,20 @@ const PlayerShareMenu: React.FC<IPlayerShareMenuProps> = ({
 
           <CopyLinkButton />
 
-          <FileDownloadButton />
+          {!isPrivate && (
+            <>
+              <FileDownloadButton />
 
-          {embedHtml && (
-            <MenuButton
-              action="clipboard"
-              clipboardText={embedHtml}
-              label="Embed Code"
-            >
-              <CodeIcon />
-            </MenuButton>
+              {embedHtml && (
+                <MenuButton
+                  action="clipboard"
+                  clipboardText={embedHtml}
+                  label="Embed Code"
+                >
+                  <CodeIcon />
+                </MenuButton>
+              )}
+            </>
           )}
         </nav>
       </Modal>
