+++ interfaces/embed/IEmbed.ts
@@ -1,0 +1,7 @@
+import { IEmbedConfig } from '@interfaces/config';
+import { IEmbedData } from '@interfaces/data';
+
+export interface IEmbedProps {
+  config: IEmbedConfig;
+  data: IEmbedData;
+}
