+++ interfaces/page/index.ts
@@ -1,3 +1,4 @@
 export * from './IEmbedPageProps';
 export * from './IListenPageProps';
 export * from './IPageProps';
+export * from './IPreviewPageProps';
